+++ gnsq/__init__.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import
 
 from .consumer import Consumer
 from .reader import Reader
+from .producer import Producer
 from .nsqd import Nsqd, NsqdTCPClient, NsqdHTTPClient
 from .lookupd import Lookupd, LookupdClient
 from .message import Message
@@ -16,6 +17,7 @@ __version__ = __version__
 __all__ = [
     'Consumer',
     'Reader',
+    'Producer',
     'Nsqd',
     'NsqdTCPClient',
     'NsqdHTTPClient',